+++ NEWS.md
@@ -1,4 +1,10 @@
-# fusen (development version)
+# fusen 0.6.0
+
+## Breaking changes
+
+- {fusen} now relies on {lightparser} instead of {parsermd} to parse flat file. This allows to avoid installation problems with {parsermd}, which is not updated anymore. As {lightparser} is lighter, this may have unattended effects on specific flat file cases. Please report any issue you may encounter. (#233)
+
+## New features
 
 - Allow `organisation` in `init_share_on_github()` to send to a GitHub organisation
 - Fix `load_flat_functions()` to work with VSCode